codes/requirements.txt
# TCMPipe: Traditional Chinese Medicine Pipeline Requirements
# ========================================================

# Core dependencies for TCM RAG system
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.21.0
pathlib>=1.0.1
ollama>=0.1.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0

# Machine learning and transformers
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0
modelscope>=1.9.0

# Data processing
pandas>=1.5.0
tqdm>=4.64.0
scikit-learn>=1.2.0

# Evaluation and metrics
nltk>=3.8.0
scipy>=1.9.0
pyahocorasick>=2.0.0

# API clients
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0

# Utilities
requests>=2.28.0
tiktoken>=0.4.0
jsonlines>=3.1.0

# Optional: Flash attention for better performance
# flash-attn>=2.0.0 